refactor: separate supervisor generation owners
